removed `codonvarianttable`
Instead we now recommend using the separate
`dms_variants` package, which defines a
`CodonVariantTable`'
