+++ dms_tools2/barcodes.py
@@ -27,9 +27,6 @@ import dms_tools2.utils
 import dms_tools2.pacbio
 from dms_tools2 import CODON_TO_AA, CODONS, AAS_WITHSTOP, AA_TO_CODONS
 
-# import CodonVariantTable and tidy_split for backward compatibility
-from dms_tools2.codonvarianttable import CodonVariantTable, tidy_split
-
 _umi_clusterer = umi_tools.network.UMIClusterer()
 
 
@@ -931,6 +928,43 @@ class IlluminaBarcodeParser:
                  )
 
         return (barcodes, fates)
+
+
+def tidy_split(df, column, sep=' ', keep=False):
+    """
+    Split values of a column and expand so new DataFrame has one split
+    value per row. Filters rows where the column is missing.
+
+    Taken from https://stackoverflow.com/a/39946744
+
+    Args:
+        df : pandas DataFrame
+            dataframe with the column to split and expand
+        column : str
+            the column to split and expand
+        sep : str
+            the string used to split the column's values
+        keep : bool
+            whether to retain the presplit value as it's own row
+
+    Returns:
+        pandas DataFrame
+            Returns a dataframe with the same columns as `df`.
+    """
+    indexes = list()
+    new_values = list()
+    df = df.dropna(subset=[column])
+    for i, presplit in enumerate(df[column].astype(str)):
+        values = presplit.split(sep)
+        if keep and len(values) > 1:
+            indexes.append(i)
+            new_values.append(presplit)
+        for value in values:
+            indexes.append(i)
+            new_values.append(value)
+    new_df = df.iloc[indexes, :].copy()
+    new_df[column] = new_values
+    return new_df
 
 
 if __name__ == '__main__':
